+++ Elements/src/Search/Node.cs
@@ -4,22 +4,22 @@ namespace Elements.Search
     /// A generic binary tree node.
     /// </summary>
     /// <typeparam name="T">The type of the data referenced by the node.</typeparam>
-    public class Node<T>
+    public class BinaryTreeNode<T>
     {
         /// <summary>
         /// The left child.
         /// </summary>
-        public Node<T> Left { get; set; }
+        public BinaryTreeNode<T> Left { get; set; }
 
         /// <summary>
         /// The right child.
         /// </summary>
-        public Node<T> Right { get; set; }
+        public BinaryTreeNode<T> Right { get; set; }
 
         /// <summary>
         /// The parent.
         /// </summary>
-        public Node<T> Parent { get; set; }
+        public BinaryTreeNode<T> Parent { get; set; }
 
         /// <summary>
         /// The data to store.